refactor(domain): extract business logic into domain package
