+++ packages/domain/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@anilog/api",
+  "name": "@anilog/domain",
   "version": "0.1.0",
   "type": "module",
   "main": "./src/index.ts",
@@ -16,12 +16,10 @@
     "@anilog/config": "workspace:*",
     "@anilog/contracts": "workspace:*",
     "@anilog/db": "workspace:*",
-    "@anilog/domain": "workspace:*",
-    "drizzle-orm": "^0.45.1",
-    "elysia": "^1.4.26"
+    "drizzle-orm": "^0.45.1"
   },
   "devDependencies": {
-    "@types/bun": "latest",
-    "typescript": "^5.6.3"
+    "@types/bun": "catalog:",
+    "typescript": "catalog:"
   }
 }
